Fix failing test for multiple quoted tags
The regular expression that matches quoted tags matches a delimiter at the start
and at the end. Because of this, when matching multiple consecutive quoted tags,
the delimiter in between is already matched as end-delimiter for the first tag,
and so can no longer serve as start-delimiter for the second tag.

By using zero-width lookahead in the regexp, the second delimiter is not
consumed.

Conflicts:
	lib/acts_as_taggable_on/tag_list.rb

bf4 - Modified and annotated expanded regex
Closes #394