adjusted specs to account for the unique index on tag.name.
Some case sensitive specs are not able to run on mysql because the default is case insensitive collation.  This combined with a unique index makes adding tags Awesome and awesome impossible because the database sees them as the same thing.  To make some of the case sensitive spec pass on mysql, I had to remove the index just for those specs.  I added a shared_example that is included where needed.