Remove existing selects from relation
If there are existing, non-standard selects on a relation, the primary key sub-select will fail because it will include them as well as the primary key, which results in too many columns being present in the results.

This change removes any existing select clauses before applying the selection on the primary key.